id->long. lock & gateway module.

+++ door/src/main/java/com/dnk/clever/door/dao/BuildDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface BuildDao extends CommonsDao<Build> {
 
-	List<Build> findList(@Param("name") String name, @Param("code") Integer code, @Param("offset") int offset, @Param("limit") int limit);
+	List<Build> findList(@Param("name") String name, @Param("code") Long code, @Param("offset") int offset, @Param("limit") int limit);
 
-	int countList(@Param("name") String name, @Param("code") Integer code);
+	int countList(@Param("name") String name, @Param("code") Long code);
 }
