+++ door/src/main/java/com/dnk/clever/door/dao/HouseDao.java
@@ -8,9 +8,9 @@ import java.util.Map;
 
 public interface HouseDao extends CommonsDao<House> {
 
-	List<House> findList(@Param("unitId") Integer unitId, @Param("name") String name, @Param("code") Integer code, @Param("offset") int offset, @Param("limit") int limit);
+	List<House> findList(@Param("unitId") Long unitId, @Param("name") String name, @Param("code") Long code, @Param("offset") int offset, @Param("limit") int limit);
 
-	List<Map> findMap(@Param("build") String build, @Param("unit") String unit, @Param("house") String house, @Param("code") Integer code, @Param("offset") int offset, @Param("limit") int limit);
+	List<Map> findMap(@Param("build") String build, @Param("unit") String unit, @Param("house") String house, @Param("code") Long code, @Param("offset") int offset, @Param("limit") int limit);
 
-	int countMap(@Param("build") String build, @Param("unit") String unit, @Param("house") String house, @Param("code") Integer code);
+	int countMap(@Param("build") String build, @Param("unit") String unit, @Param("house") String house, @Param("code") Long code);
 }
