+++ door/src/main/java/com/dnk/clever/door/dao/UnitDao.java
@@ -8,10 +8,10 @@ import java.util.Map;
 
 public interface UnitDao extends CommonsDao<Unit> {
 
-	List<Unit> findList(@Param("buildId") Integer buildId, @Param("name") String name, @Param("code") Integer code, @Param("offset") int offset, @Param("limit") int limit);
+	List<Unit> findList(@Param("buildId") Long buildId, @Param("name") String name, @Param("code") Long code, @Param("offset") int offset, @Param("limit") int limit);
 
-	List<Map> findMap(@Param("build") String build, @Param("unit") String unit, @Param("code") Integer code, @Param("offset") int offset, @Param("limit") int limit);
+	List<Map> findMap(@Param("build") String build, @Param("unit") String unit, @Param("code") Long code, @Param("offset") int offset, @Param("limit") int limit);
 
-	int countMap(@Param("build") String build, @Param("unit") String unit, @Param("code") Integer code);
+	int countMap(@Param("build") String build, @Param("unit") String unit, @Param("code") Long code);
 
 }
