+++ door/src/main/java/com/dnk/clever/door/service/CommonsService.java
@@ -12,11 +12,11 @@ public interface CommonsService<Entity> {
 
 	int saves(Collection<Entity> entities);
 
-	int deleteById(int id);
+	int deleteById(long id);
 
 	int deleteByEntity(Entity entity);
 
-	int deleteByIds(int[] ids);
+	int deleteByIds(long[] ids);
 
 	int deleteByEntities(Collection<Entity> entities);
 
@@ -27,7 +27,7 @@ public interface CommonsService<Entity> {
 		return entities.size();
 	}
 
-	Entity findById(int id);
+	Entity findById(long id);
 
 	List<Entity> findAll();
